refactor: rename S3 bucket, relocate tests, update deps
- Rename ICEBERG_WAREHOUSE bucket from dimo-iceberg-* to dimo-storage-*
  across stream configs and Helm values (dev + prod)
- Fix import path: cloudeventparquet/encoders -> encoders
- Move parseBucket tests into cloudeventtoparquet package and
  RawEvent_DataBase64 test into encoders; delete parquetwriter_test.go
- Update go.mod: add aws-sdk-go-v2 core + s3, drop segmentio/encoding,
  pin tidwall/match v1.1.1
- Fix struct field alignment in icebergcommit_test.go
