+++ charts/dps/values-prod.yaml
@@ -10,7 +10,7 @@ env:
   LOG_LEVEL: INFO
   # Iceberg / Object Storage (cloud-agnostic)
   ICEBERG_CATALOG_URI: http://lakekeeper:8181
-  ICEBERG_WAREHOUSE: s3://dimo-iceberg-prod/warehouse/
+  ICEBERG_WAREHOUSE: s3://dimo-storage-prod/warehouse/
   OBJECT_STORAGE_ENDPOINT: https://s3.us-east-2.amazonaws.com
 ingress:
   enabled: false
