+++ internal/processors/cloudeventtoparquet/cloudeventtoparquet_test.go
@@ -5,8 +5,36 @@ import (
 	"testing"
 
 	"github.com/redpanda-data/benthos/v4/public/service"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
+
+func TestParseBucket(t *testing.T) {
+	tests := []struct {
+		name      string
+		warehouse string
+		expected  string
+		wantErr   bool
+	}{
+		{name: "standard", warehouse: "s3://my-bucket/warehouse/", expected: "my-bucket"},
+		{name: "no trailing slash", warehouse: "s3://my-bucket", expected: "my-bucket"},
+		{name: "with path", warehouse: "s3://dimo-storage-dev/warehouse/data/", expected: "dimo-storage-dev"},
+		{name: "invalid prefix", warehouse: "gs://bucket/path/", wantErr: true},
+		{name: "empty", warehouse: "", wantErr: true},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := parseBucket(tt.warehouse)
+			if tt.wantErr {
+				require.Error(t, err)
+				return
+			}
+			require.NoError(t, err)
+			assert.Equal(t, tt.expected, got)
+		})
+	}
+}
 
 func TestProcessBatch_MetadataAndBucket(t *testing.T) {
 	mgr := service.MockResources()
